LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/query-evaluation-expected.txt


PASS size(width)
FAIL size(height) assert_equals: expected "false" but got "true"
FAIL size(unknown) assert_equals: expected "false" but got "true"
FAIL unknown(width) assert_equals: expected "false" but got "true"
PASS (size(width))
FAIL (size(height)) assert_equals: expected "false" but got "true"
FAIL (size(unknown)) assert_equals: expected "false" but got "true"
PASS (((size(width))))
FAIL (((size(height)))) assert_equals: expected "false" but got "true"
FAIL (((size(unknown)))) assert_equals: expected "false" but got "true"
FAIL (not size(width)) assert_equals: expected "false" but got "true"
PASS (not size(height))
FAIL (not size(unknown)) assert_equals: expected "false" but got "true"
FAIL (not unknown(width)) assert_equals: expected "false" but got "true"
PASS (size(width) and size(width))
PASS (size(width) and size(width) and size(width))
FAIL (size(height) and size(height)) assert_equals: expected "false" but got "true"
FAIL (size(height) and size(width) and size(width)) assert_equals: expected "false" but got "true"
FAIL (size(width) and size(height) and size(width)) assert_equals: expected "false" but got "true"
FAIL (size(width) and size(width) and size(height)) assert_equals: expected "false" but got "true"
FAIL (size(unknown) and size(width) and size(width)) assert_equals: expected "false" but got "true"
FAIL (size(width) and size(unknown) and size(width)) assert_equals: expected "false" but got "true"
FAIL (size(width) and size(width) and size(unknown)) assert_equals: expected "false" but got "true"
PASS (size(width) or size(width))
PASS (size(width) or size(width) or size(width))
FAIL (size(height) or size(height)) assert_equals: expected "false" but got "true"
PASS (size(height) or size(width) or size(width))
PASS (size(width) or size(height) or size(width))
PASS (size(width) or size(width) or size(height))
PASS (size(unknown) or size(width) or size(width))
PASS (size(width) or size(unknown) or size(width))
PASS (size(width) or size(width) or size(unknown))
PASS (size(unknown) or size(height) or size(width))
FAIL (not (size(width) and size(width))) assert_equals: expected "false" but got "true"
PASS (not (size(width) and size(height)))
FAIL (size(width) and (not (size(height) or size(width)))) assert_equals: expected "false" but got "true"
PASS (size(height) or (not (size(height) and size(width))))
FAIL (size(height) or (size(height) and size(width))) assert_equals: expected "false" but got "true"
PASS size((width))
FAIL size((height)) assert_equals: expected "false" but got "true"
FAIL size((unknown)) assert_equals: expected "false" but got "true"
FAIL unknown((width)) assert_equals: expected "false" but got "true"
FAIL size(not (width)) assert_equals: expected "false" but got "true"
PASS size(not (height))
FAIL size(not (unknown)) assert_equals: expected "false" but got "true"
PASS size((width) and (width))
PASS size((width) and (width) and (width))
FAIL size((height) and (height)) assert_equals: expected "false" but got "true"
FAIL size((height) and (width) and (width)) assert_equals: expected "false" but got "true"
FAIL size((width) and (height) and (width)) assert_equals: expected "false" but got "true"
FAIL size((width) and (width) and (height)) assert_equals: expected "false" but got "true"
FAIL size((unknown) and (width) and (width)) assert_equals: expected "false" but got "true"
FAIL size((width) and (unknown) and (width)) assert_equals: expected "false" but got "true"
FAIL size((width) and (width) and (unknown)) assert_equals: expected "false" but got "true"
PASS size((width) or (width))
PASS size((width) or (width) or (width))
FAIL size((height) or (height)) assert_equals: expected "false" but got "true"
PASS size((height) or (width) or (width))
PASS size((width) or (height) or (width))
PASS size((width) or (width) or (height))
PASS size((unknown) or (width) or (width))
PASS size((width) or (unknown) or (width))
PASS size((width) or (width) or (unknown))
PASS size((unknown) or (height) or (width))
FAIL size(not ((width) and (width))) assert_equals: expected "false" but got "true"
PASS size(not ((width) and (height)))
FAIL size((width) and (not ((height) or (width)))) assert_equals: expected "false" but got "true"
PASS size((height) or (not ((height) and (width))))
FAIL size((height) or ((height) and (width))) assert_equals: expected "false" but got "true"
